+++ eduVPN.Views/App.cs
@@ -23,6 +23,15 @@ namespace eduVPN.Views
     /// </summary>
     public class App : Application, ISingleInstanceApp
     {
+        #region Properties
+
+        /// <summary>
+        /// Was client started at user sign-on?
+        /// </summary>
+        public bool IsSignon { get; private set; }
+
+        #endregion
+
         #region Methods
 
         /// <inheritdoc/>
@@ -66,12 +75,16 @@ namespace eduVPN.Views
                 else
                 {
                     // User cancelled. Quit.
-                    return;
+                    Shutdown(1);
                 }
             }
 
             eduVPN.Properties.Settings.Initialize();
             Views.Properties.Settings.Initialize();
+
+            IsSignon = e.Args.Any(param => param.Equals("/signon", StringComparison.OrdinalIgnoreCase));
+            if (IsSignon && !Views.Properties.Settings.Default.StartOnSignon)
+                Shutdown(0);
 
             RenderOptions.ProcessRenderMode = Views.Properties.SettingsEx.Default.ProcessRenderMode;